requirements.txt
streamlit>=1.28.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
python-dotenv>=1.0.0
cryptography>=41.0.0
pandas>=2.1.0
pycryptodome>=3.19.0 